Implement working solution for mapInput on native
mapInput does not work on native. 
On native, you cannot catch the error when calling map with the wrong input type.
Rework by passing the kclass to mapInput

Store mappers in map of KClass to Mapper
Also in the mapper interface add inputKClass and outputKClass.
inputKClass is used to store a mapper in the new mappers (KClass to Mapper) map in KamelConfigBuilder