Output cached result immediately, if available.
Currently the `lazyPainterResource` will always result in at least 2
compositions even when the painter is already cached: the first one
always returns Resource.Loading, while the second one is the actual
Resource.Success containing the cached Painter.

By looking up the cached Painter ahead of time and using that as the
initial value if it exists, we skip a recomposition for preloaded
Painters and avoid a UI flash in some cases where an external factor
causes a recomposition.